worker/requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP client
httpx>=0.26.0

# Transcription
faster-whisper>=1.0.3

# Data handling
numpy>=1.26.0
pydantic>=2.5.0

# OCR and document processing
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# Utilities
python-slugify>=8.0.0

# AWQ quantization support
autoawq>=0.2.0
